Move NotFound404 & test file to its own folder
